[FIX] 시나리오 생성시 response버그문제해결
