+++ back/src/test/java/com/back/domain/scenario/service/ScenarioServiceTest.java
@@ -44,10 +44,9 @@ class ScenarioServiceTest {
     @Mock private ScenarioRepository scenarioRepository;
     @Mock private DecisionLineRepository decisionLineRepository;
     @Mock private ObjectMapper objectMapper;
-    @Mock private AiService aiService;
     @Mock private ScenarioTransactionService scenarioTransactionService;
 
-    @Spy @InjectMocks private ScenarioService scenarioService;
+    @InjectMocks private ScenarioService scenarioService;
 
     @Nested
     @DisplayName("시나리오 생성")
@@ -89,7 +88,7 @@ class ScenarioServiceTest {
                     .willReturn(savedScenario);
 
             // 비동기 메서드 호출을 무효화 (동기 로직만 테스트)
-            doNothing().when(scenarioService).processScenarioGenerationAsync(anyLong());
+            doNothing().when(scenarioTransactionService).processScenarioGenerationAsync(anyLong());
 
             // When
             ScenarioStatusResponse result = scenarioService.createScenario(userId, request, null);
@@ -105,8 +104,8 @@ class ScenarioServiceTest {
             verify(scenarioRepository).findByDecisionLineId(decisionLineId);
             verify(scenarioRepository).save(any(Scenario.class));
 
-            // 비동기 메서드가 호출되었는지 확인 (하지만 실행은 되지 않음)
-            verify(scenarioService).processScenarioGenerationAsync(1001L);
+            // 비동기 메서드가 호출되었는지 확인 (ScenarioTransactionService를 통해)
+            verify(scenarioTransactionService).processScenarioGenerationAsync(1001L);
         }
 
         @Test
@@ -193,92 +192,6 @@ class ScenarioServiceTest {
                     .hasFieldOrPropertyWithValue("errorCode", ErrorCode.SCENARIO_ALREADY_IN_PROGRESS);
 
             verify(scenarioRepository, never()).save(any());
-        }
-    }
-
-    @Nested
-    @DisplayName("비동기 시나리오 생성 처리")
-    class ProcessScenarioGenerationAsyncTests {
-
-        @Test
-        @DisplayName("성공 - 비동기 시나리오 생성 및 완료")
-        void processScenarioGenerationAsync_성공_시나리오생성완료() {
-            // Given
-            Long scenarioId = 1001L;
-            User mockUser = User.builder().build();
-            ReflectionTestUtils.setField(mockUser, "id", 1L);
-
-            BaseLine mockBaseLine = BaseLine.builder().user(mockUser).build();
-            ReflectionTestUtils.setField(mockBaseLine, "id", 100L); // BaseLine ID 설정
-
-            DecisionLine mockDecisionLine = DecisionLine.builder()
-                    .user(mockUser)
-                    .baseLine(mockBaseLine)
-                    .build();
-            ReflectionTestUtils.setField(mockDecisionLine, "id", 200L);
-
-            Scenario mockScenario = Scenario.builder()
-                    .user(mockUser)
-                    .decisionLine(mockDecisionLine)
-                    .baseLine(mockBaseLine)
-                    .status(ScenarioStatus.PENDING)
-                    .build();
-            ReflectionTestUtils.setField(mockScenario, "id", scenarioId);
-
-            // 트랜잭션 분리된 메서드들 모킹
-            doNothing().when(scenarioTransactionService).updateScenarioStatus(anyLong(), any(), any());
-            doNothing().when(scenarioTransactionService).saveAiResult(anyLong(), any());
-
-            // prepareScenarioData 모킹 추가 (핵심!)
-            given(scenarioTransactionService.prepareScenarioData(scenarioId))
-                    .willReturn(mockScenario);
-
-            // 베이스 시나리오가 이미 존재하도록 설정 (AI 호출 방지)
-            Scenario mockBaseScenario = Scenario.builder()
-                    .user(mockUser)
-                    .baseLine(mockBaseLine)
-                    .status(ScenarioStatus.COMPLETED)
-                    .build();
-            ReflectionTestUtils.setField(mockBaseScenario, "id", 999L);
-
-            given(scenarioRepository.findByBaseLineIdAndDecisionLineIsNull(any()))
-                    .willReturn(Optional.of(mockBaseScenario)); // 베이스 시나리오 이미 존재
-
-            // DecisionScenario 생성용 AI 서비스 모킹
-            given(aiService.generateDecisionScenario(any(DecisionLine.class), any(Scenario.class)))
-                    .willReturn(CompletableFuture.completedFuture(mockDecisionScenarioResult()));
-
-            // When
-            scenarioService.processScenarioGenerationAsync(scenarioId);
-
-            // Then
-            // 트랜잭션 서비스 호출 검증
-            verify(scenarioTransactionService).updateScenarioStatus(scenarioId, ScenarioStatus.PROCESSING, null);
-            verify(scenarioTransactionService).prepareScenarioData(scenarioId);
-            verify(scenarioTransactionService).updateScenarioStatus(scenarioId, ScenarioStatus.COMPLETED, null);
-        }
-
-        @Test
-        @DisplayName("실패 - 존재하지 않는 시나리오 (비동기 처리)")
-        void processScenarioGenerationAsync_실패_존재하지않는_시나리오() {
-            // Given
-            Long scenarioId = 999L;
-
-            // 트랜잭션 서비스는 정상 동작하도록 모킹
-            doNothing().when(scenarioTransactionService).updateScenarioStatus(anyLong(), any(), any());
-
-            // prepareScenarioData에서 시나리오를 찾을 수 없음 (예외 발생)
-            given(scenarioTransactionService.prepareScenarioData(scenarioId))
-                    .willThrow(new ApiException(ErrorCode.SCENARIO_NOT_FOUND));
-
-            // When
-            scenarioService.processScenarioGenerationAsync(scenarioId);
-
-            // Then
-            // 실패 시 FAILED 상태 업데이트가 호출되어야 함
-            verify(scenarioTransactionService).updateScenarioStatus(scenarioId, ScenarioStatus.PROCESSING, null);
-            verify(scenarioTransactionService).prepareScenarioData(scenarioId);
-            verify(scenarioTransactionService).updateScenarioStatus(eq(scenarioId), eq(ScenarioStatus.FAILED), anyString());
         }
     }
 
@@ -469,24 +382,5 @@ class ScenarioServiceTest {
             ScenarioStatusResponse failedResult = scenarioService.getScenarioStatus(scenarioId, userId);
             assertThat(failedResult.message()).isEqualTo("시나리오 생성에 실패했습니다. 다시 시도해주세요.");
         }
-    }
-
-    // 헬퍼 메서드: Mock DecisionScenario 결과 생성
-    private DecisionScenarioResult mockDecisionScenarioResult() {
-        return new DecisionScenarioResult(
-                "테스트 직업",
-                "테스트 요약",
-                "테스트 설명",
-                100, // total
-                "테스트 이미지",
-                Map.of("2025", "테스트 타이틀"),
-                java.util.List.of(
-                        new DecisionScenarioResult.Indicator("경제", 50, "테스트 경제 분석"),
-                        new DecisionScenarioResult.Indicator("행복", 50, "테스트 행복 분석")
-                ),
-                java.util.List.of(
-                        new DecisionScenarioResult.Comparison("TOTAL", 200, 250, "테스트 비교 분석")
-                )
-        );
     }
 }
